// ClientApp_old/src/services/openapi/api.ts
export interface AddFoodDiaryRequest {
    /**
     * 
     * @type {string}
     * @memberof AddFoodDiaryRequest
     */
    'foodId': string;
    /**
     * 
     * @type {MealTypes}
     * @memberof AddFoodDiaryRequest
     */
    'mealType': MealTypes;
    /**
     * 
     * @type {string}
     * @memberof AddFoodDiaryRequest
     */
    'servingSizeId': string;
    /**
     * 
     * @type {number}
     * @memberof AddFoodDiaryRequest
     */
    'quantity': number;
    /**
     * 
     * @type {string}
     * @memberof AddFoodDiaryRequest
     */
    'entryDate': string;
}
// ...
export interface AddRecipeDiaryRequest {
    /**
     * 
     * @type {string}
     * @memberof AddRecipeDiaryRequest
     */
    'recipeId': string;
    /**
     * 
     * @type {MealTypes}
     * @memberof AddRecipeDiaryRequest
     */
    'mealType': MealTypes;
    /**
     * 
     * @type {number}
     * @memberof AddRecipeDiaryRequest
     */
    'quantity': number;
    /**
     * 
     * @type {string}
     * @memberof AddRecipeDiaryRequest
     */
    'entryDate': string;
}


/**
 * 
 * @export
 * @interface CreateRecipeRequest
 */
export interface CreateRecipeRequest {
    /**
     * 
     * @type {string}
     * @memberof CreateRecipeRequest
     */
    'name': string;
}
/**
 * 
 * @export
 * @enum {string}
 */

export const DiaryType = {
    FoodDiary: 'FoodDiary',
    RecipeDiary: 'RecipeDiary'
} as const;

export type DiaryType = typeof DiaryType[keyof typeof DiaryType];


/**
 * 
 * @export
 * @interface Energy
 */
export interface Energy {
    /**
     * 
     * @type {string}
     * @memberof Energy
     */
    'unit': string;
    /**
     * 
     * @type {number}
     * @memberof Energy
     */
    'value': number;
}
// ...
export const MealTypes = {
    Breakfast: 'Breakfast',
    Lunch: 'Lunch',
    Dinner: 'Dinner',
    Snacks: 'Snacks'
} as const;

export type MealTypes = typeof MealTypes[keyof typeof MealTypes];


/**
 * 
 * @export
 * @interface NutritionDiaryDto
 */
export interface NutritionDiaryDto {
    /**
     * 
     * @type {string}
     * @memberof NutritionDiaryDto
     */
    'id': string;
    /**
     * 
     * @type {string}
     * @memberof NutritionDiaryDto
     */
    'name': string;
    /**
     * 
     * @type {NutritionDiaryDtoServingSize}
     * @memberof NutritionDiaryDto
     */
    'servingSize'?: NutritionDiaryDtoServingSize | null;
    /**
     * 
     * @type {NutritionalContent}
     * @memberof NutritionDiaryDto
     */
    'nutritionalContents': NutritionalContent;
    /**
     * 
     * @type {number}
     * @memberof NutritionDiaryDto
     */
    'nutritionMultiplier': number;
    /**
     * 
     * @type {number}
     * @memberof NutritionDiaryDto
     */
    'quantity': number;
    /**
     * 
     * @type {MealTypes}
     * @memberof NutritionDiaryDto
     */
    'mealType': MealTypes;
    /**
     * 
     * @type {DiaryType}
     * @memberof NutritionDiaryDto
     */
    'diaryType': DiaryType;
    /**
     * 
     * @type {string}
     * @memberof NutritionDiaryDto
     */
    'date': string;
}


/**
 * @type NutritionDiaryDtoServingSize
 * @export
 */
export type NutritionDiaryDtoServingSize = ServingSizeDto;

/**
 * 
 * @export
 * @interface NutritionalContent
 */
export interface NutritionalContent {
    /**
     * 
     * @type {number}
     * @memberof NutritionalContent
     */
    'calcium': number;
    /**
     * 
     * @type {number}
     * @memberof NutritionalContent
     */
    'carbohydrates': number;
    /**
     * 
     * @type {number}
     * @memberof NutritionalContent
     */
    'cholesterol': number;
    /**
     * 
     * @type {number}
     * @memberof NutritionalContent
     */
    'fat': number;
    /**
     * 
     * @type {number}
     * @memberof NutritionalContent
     */
    'fiber': number;
    /**
     * 
     * @type {number}
     * @memberof NutritionalContent
     */
    'iron': number;
    /**
     * 
     * @type {number}
     * @memberof NutritionalContent
     */
    'monounsaturatedFat': number;
    /**
     * 
     * @type {number}
     * @memberof NutritionalContent
     */
    'netCarbs': number;
    /**
     * 
     * @type {number}
     * @memberof NutritionalContent
     */
    'polyunsaturatedFat': number;
    /**
     * 
     * @type {number}
     * @memberof NutritionalContent
     */
    'potassium': number;
    /**
     * 
     * @type {number}
     * @memberof NutritionalContent
     */
    'protein': number;
    /**
     * 
     * @type {number}
     * @memberof NutritionalContent
     */
    'saturatedFat': number;
    /**
     * 
     * @type {number}
     * @memberof NutritionalContent
     */
    'sodium': number;
    /**
     * 
     * @type {number}
     * @memberof NutritionalContent
     */
    'sugar': number;
    /**
     * 
     * @type {number}
     * @memberof NutritionalContent
     */
    'transFat': number;
    /**
     * 
     * @type {number}
     * @memberof NutritionalContent
     */
    'vitaminA': number;
    /**
     * 
     * @type {number}
     * @memberof NutritionalContent
     */
    'vitaminC': number;
    /**
     * 
     * @type {Energy}
     * @memberof NutritionalContent
     */
    'energy': Energy;
}
// ...
export interface ServingSizeDto {
    /**
     * 
     * @type {string}
     * @memberof ServingSizeDto
     */
    'id': string;
    /**
     * 
     * @type {number}
     * @memberof ServingSizeDto
     */
    'nutritionMultiplier': number;
    /**
     * 
     * @type {number}
     * @memberof ServingSizeDto
     */
    'value': number;
    /**
     * 
     * @type {string}
     * @memberof ServingSizeDto
     */
    'unit': string;
}
// ...
export interface UpdateFoodDiaryRequest {
    /**
     * 
     * @type {string}
     * @memberof UpdateFoodDiaryRequest
     */
    'id': string;
    /**
     * 
     * @type {number}
     * @memberof UpdateFoodDiaryRequest
     */
    'quantity': number;
    /**
     * 
     * @type {string}
     * @memberof UpdateFoodDiaryRequest
     */
    'servingSizeId': string;
    /**
     * 
     * @type {MealTypes}
     * @memberof UpdateFoodDiaryRequest
     */
    'mealType': MealTypes;
}
// ...
export interface UpdateRecipeDiaryRequest {
    /**
     * 
     * @type {string}
     * @memberof UpdateRecipeDiaryRequest
     */
    'id': string;
    /**
     * 
     * @type {number}
     * @memberof UpdateRecipeDiaryRequest
     */
    'quantity': number;
    /**
     * 
     * @type {MealTypes}
     * @memberof UpdateRecipeDiaryRequest
     */
    'mealType': MealTypes;
}
// ...
export class FoodDiariesApi extends BaseAPI {
    /**
     * 
     * @param {AddFoodDiaryRequest} addFoodDiaryRequest 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof FoodDiariesApi
     */
    public addFoodDiary(addFoodDiaryRequest: AddFoodDiaryRequest, options?: RawAxiosRequestConfig) {
        return FoodDiariesApiFp(this.configuration).addFoodDiary(addFoodDiaryRequest, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {string | null} id 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof FoodDiariesApi
     */
    public deleteFoodDiary(id: string | null, options?: RawAxiosRequestConfig) {
        return FoodDiariesApiFp(this.configuration).deleteFoodDiary(id, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {string | null} id 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof FoodDiariesApi
     */
    public getFoodDiaryById(id: string | null, options?: RawAxiosRequestConfig) {
        return FoodDiariesApiFp(this.configuration).getFoodDiaryById(id, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {UpdateFoodDiaryRequest} updateFoodDiaryRequest 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof FoodDiariesApi
     */
    public updateFoodDiary(updateFoodDiaryRequest: UpdateFoodDiaryRequest, options?: RawAxiosRequestConfig) {
        return FoodDiariesApiFp(this.configuration).updateFoodDiary(updateFoodDiaryRequest, options).then((request) => request(this.axios, this.basePath));
    }
}
// ...
export class RecipeDiariesApi extends BaseAPI {
    /**
     * 
     * @param {AddRecipeDiaryRequest} addRecipeDiaryRequest 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof RecipeDiariesApi
     */
    public addRecipeDiary(addRecipeDiaryRequest: AddRecipeDiaryRequest, options?: RawAxiosRequestConfig) {
        return RecipeDiariesApiFp(this.configuration).addRecipeDiary(addRecipeDiaryRequest, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {string | null} id 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof RecipeDiariesApi
     */
    public deleteRecipeDiary(id: string | null, options?: RawAxiosRequestConfig) {
        return RecipeDiariesApiFp(this.configuration).deleteRecipeDiary(id, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {string | null} id 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof RecipeDiariesApi
     */
    public getRecipeDiaryById(id: string | null, options?: RawAxiosRequestConfig) {
        return RecipeDiariesApiFp(this.configuration).getRecipeDiaryById(id, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {UpdateRecipeDiaryRequest} updateRecipeDiaryRequest 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof RecipeDiariesApi
     */
    public updateRecipeDiary(updateRecipeDiaryRequest: UpdateRecipeDiaryRequest, options?: RawAxiosRequestConfig) {
        return RecipeDiariesApiFp(this.configuration).updateRecipeDiary(updateRecipeDiaryRequest, options).then((request) => request(this.axios, this.basePath));
    }
}